// frontend/src/generated/openapi/model/detallesLibro.ts
/**
 * amoxtli
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Información detallada respecto a un libro
 */
export interface DetallesLibro { 
    /**
     * ISBN (International Standard Book Number) del libro
     */
    isbn: string;
    /**
     * Autor del libro
     */
    autor: string;
    /**
     * Título del libro
     */
    titulo: string;
    /**
     * URL de la portada del libro
     */
    urlPortada: string;
    /**
     * Géneros literarios del libro
     */
    generos: Array<string>;
    /**
     * Editorial que publica el libro
     */
    editorial: string;
    /**
     * Sinopsis del libro
     */
    sinopsis?: string;
    /**
     * Idioma del libro
     */
    idioma: string;
    /**
     * Fecha de publicación del libro
     */
    fechaPublicacion: string;
}