// frontend/src/generated/openapi/model/modelError.ts
/**
 * amoxtli
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Información de un error
 */
export interface ModelError { 
    /**
     * Mensaje amigable que describe el error
     */
    mensaje: string;
    /**
     * Código único asociado con el error
     */
    codigo: number;
}